changed module structure
