+++ src/app/components/anime/add-anime/add-anime.component.ts
@@ -9,9 +9,9 @@ import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { take } from 'rxjs/operators';
 
-import { IAddEditAnime } from 'src/app/models/anime.mode';
+import { IAddEditAnime } from 'src/app/components/anime/anime.mode';
 
-import { AnimeService } from '../../../services/anime/anime.service';
+import { AnimeService } from '../service/anime.service';
 
 @Component({
   selector: 'app-add-anime',
