+++ src/app/components/heroes/add-edit-hero/add-edit-hero.component.ts
@@ -5,12 +5,12 @@ import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 
 import { mimeTypeValidator } from 'src/app/help-functions/image.validator';
 
-import { AnimeService } from '../../../services/anime/anime.service';
-import { HeroesService } from '../../../services/heroes/heroes.service';
+import { AnimeService } from '../../anime/service/anime.service';
+import { HeroesService } from '../service/heroes.service';
 import {
   IAddEditHeroDialogData,
   IAddEditHero,
-} from 'src/app/models/hero.model';
+} from 'src/app/components/heroes/hero.model';
 
 @Component({
   selector: 'app-add-edit-hero',
