+++ src/app/components/home/home.component.scss
@@ -1,5 +1,5 @@
 .container {
-  background: url("./../../../../assets/background/home.png") 0 0/100% 100%
+  background: url("./../../../assets/background/home.png") 0 0/100% 100%
     no-repeat;
   height: 100vh;
 }
