+++ src/app/components/home/home.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+
+import { HomeComponent } from './home.component';
+
+@NgModule({
+  declarations: [HomeComponent],
+  imports: [],
+  exports: [],
+  providers: [],
+})
+export class HomeModule {}
