--- src/app/models/message-repsonse.model.ts
@@ -1,3 +1,0 @@
-export interface IMessageResponse {
-  message: string;
-}
