--- src/app/routings/heroes-routing.module.ts
@@ -1,12 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { HeroesComponent } from '../components/heroes/heroes.component';
-
-const routes: Routes = [{ path: '', component: HeroesComponent }];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class HeroesRoutingModule {}
