--- src/app/routings/home-routing.module.ts
@@ -1,12 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { HomeComponent } from '../components/home/home.component';
-
-const routes: Routes = [{ path: '', component: HomeComponent }];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class HomeRoutingModule { }
