--- src/app/routings/quotes-routing.module.ts
@@ -1,12 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { QuotesComponent } from '../components/quotes/quotes.component';
-
-const routes: Routes = [{ path: '', component: QuotesComponent }];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class QuotesRoutingModule {}
