+++ src/main.ts
@@ -1,6 +1,6 @@
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app/modules/app.module/app.module';
+import { AppModule } from './app/modules/app.module';
 
 
 platformBrowserDynamic().bootstrapModule(AppModule)
